requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
blinker==1.6.3
pytest==8.0.0
httpx==0.26.0
gunicorn==21.2.0